raspberry/requirements.txt
pyserial==3.5 
RPi.GPIO>=0.7.1
adafruit-blinka>=8.0.0
adafruit-blinka-displayio>=1.0.0
adafruit-circuitpython-st7789>=2.0.0
adafruit-circuitpython-display-text>=2.0.0
solana>=0.30.0
solders>=0.21.0
base58>=2.1.0
httpx>=0.24.0
websockets>=11.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
PyNaCl>=1.5.0